feat(agent): foldable Vue terminal with Ctrl+O toggle
Replaces the xterm.js scrollback with a Vue-native terminal that can
fold tool-call output. Tool calls collapse to a one-line summary by
default — "$ <cmd>   N lines, exit M" with green/red exit-code
colour — and Ctrl+O (or ⌘+O) toggles fold/unfold across every block at
once. Click the block header to toggle a single tool call.

Why this trade-off:
- xterm.js is a character grid; folding requires a parallel logical
  model and a full re-render on every toggle. Vue components track
  semantic blocks natively, so fold state is just reactive data.
- Side benefits: per-block text selection/copy, native theming via
  Tailwind tokens, no ANSI cursor math, no @xterm/* deps in the agent
  bundle, easier to extend (per-tool icons, inline images, links).

Implementation notes:
- store.AgentMessage gains an optional 'tool' meta field carrying
  script/stdout/stderr/exitCode so the renderer doesn't have to parse
  the synthesized text summary back into structure.
- Global keydown listener with capture:true preempts the browser's
  default Open-File dialog before Ctrl+O reaches it.
- Drops src/agent/ui/XtermPanel.vue, useXtermReadline.ts, and
  shell/useCompletion.ts. The xterm package is still used elsewhere
  in the codebase (logs/CommandTerminal), so no dep removal.
